ai-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pillow==10.3.0
numpy==1.26.4
opencv-python==4.10.0.82
imagehash==4.3.1
python-dotenv==1.0.1
requests==2.32.3
